+++ package.json
@@ -2,7 +2,7 @@
   "name": "@swarnim/hermes",
   "author": "Swarnim Walavalkar <swarnim.walavalkar@gmail.com>",
   "license": "MIT",
-  "version": "0.0.1",
+  "version": "1.1.0",
   "main": "dist/index.js",
   "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
